requirements.txt
Markdown==3.4.1
PyYAML==6.0
bleach==5.0.1
django-filter==22.1
django-formtools==2.4
django-mptt==0.14.0
django-otp-yubikey==1.0.1
django-otp==1.1.3
django-rq==2.5.1
django-tables2==2.4.1
django==4.1.2
django_browser_reload==1.6.0
djangorestframework==3.14.0
drf-yasg==1.21.4
netaddr==0.8.0
phonenumbers==8.12.57
psycopg2==2.9.5
rq-scheduler==0.11.0
segno==1.5.2
swagger-spec-validator==3.0.2
typing-extensions==4.4.0
